+++ docs/CHANGELOG.md
@@ -9,6 +9,9 @@
 * Correctly pretty-print Float16 and BFloat16 tensors.
 * Fixed nanlogger for Float16 and BFloat16; Also, it first prints other logged tensors, before failing with a NaN.
 * Several fixes and small improvements to command-line tool `gomlx_checkpoint`.
+* `nanlogger`:
+  * Store only the stack-trace, and trim the stack into the nanlogger package.
+  * Does not exit, simply report the NanLogger. User can define a handler, if they want the training to exit.
 
 # v0.16.1 - 🎄 2024/12/19 🎄 MatMul fixes
 * MatMul fixed for some edge shape configuration and greatly accelerated in some cases.